requirements.txt
numpy
matplotlib
seaborn
pillow
torch
torchvision
albumentations
fvcore
